requirements.txt
Django
djangorestframework
django-filter
django-apscheduler
gunicorn
sentence_transformers